+++ src/components/Functional/Container.tsx
@@ -1,6 +1,6 @@
 import { isMobile } from '@/ui/lib/mediaQuery';
 import { styled } from '@/ui/styled/dynamic';
-import { SPACE_STEPS } from '@/ui/styled/variables/Space';
+import { SPACE_STEPS } from '@/ui/styled/variables/space';
 
 type Props = {
   size?: 'small' | 'default' | 'large' | string;
