+++ src/components/UI/Layout/Grid.tsx
@@ -3,8 +3,8 @@ import { memo, useMemo } from 'react';
 
 import { isMobile } from '@/ui/lib/mediaQuery';
 import { styled } from '@/ui/styled/dynamic';
-import type { SpaceGap } from '@/ui/styled/variables/Space';
-import { SPACE_KEYS } from '@/ui/styled/variables/Space';
+import type { SpaceGap } from '@/ui/styled/variables/space';
+import { SPACE_KEYS } from '@/ui/styled/variables/space';
 
 type Props = {
   as?: keyof JSX.IntrinsicElements;
