+++ src/ui/styled/variables/easing.ts
